refactor: remove comentários e muda nomes
