+++ src/utils/functions.js
@@ -1,4 +1,29 @@
 const { ActionRowBuilder, ButtonBuilder } = require("discord.js")
+const fs = require("fs")
+const { promisify } = require("util")
+
+const readdir = promisify(fs.readdir)
+const stat = promisify(fs.stat)
+
+async function loadFiles(dirName) {
+	const basePath = `${process.cwd().replace(/\\/g, "/")}/src/${dirName}`
+
+	const files = []
+	const items = await readdir(basePath)
+	for (const item of items) {
+		const itemPath = `${basePath}/${item}`
+		const itemStat = await stat(itemPath)
+		if (itemStat.isDirectory()) {
+			const subFiles = await loadFiles(`${dirName}/${item}`)
+			files.push(...subFiles)
+		} else if (itemPath.endsWith(".js")) {
+			files.push(itemPath)
+			delete require.cache[require.resolve(itemPath)]
+		}
+	}
+
+	return files
+}
 
 /**
  * @param {integer} ms
@@ -37,33 +62,6 @@ function msToTime(ms) {
 	if (time === "") time += "1m"
 
 	return time.trimEnd()
-}
-
-/**
- *
- * @param {integer} falcoins
- * @description Format a number to a string with the format "1.000.000"
- * @example format(1000000) // 1.000.000
- * @returns {string}
- */
-function format(falcoins) {
-	if (parseInt(falcoins) < 0) {
-		falcoins = falcoins.toString()
-		pop = falcoins.slice(1)
-	} else {
-		pop = falcoins.toString()
-	}
-	pop_reverse = pop.split("").reverse().join("")
-	pop_2 = ""
-	for (c in pop_reverse) {
-		if (c / 3 == parseInt(c / 3) && c / 3 != 0) {
-			pop_2 += "."
-			pop_2 += pop_reverse[c]
-		} else {
-			pop_2 += pop_reverse[c]
-		}
-	}
-	return pop_2.split("").reverse().join("")
 }
 
 /**
@@ -131,33 +129,9 @@ function paginate() {
 	}
 }
 
-/**
- *
- * @param {Array<string, number>} data
- * @description Picks a random element from an array based on its weight
- * @example pick([['a', 0.5], ['b', 0.3], ['c', 0.2]]) // 'a'
- * @returns {string}
- */
-function pick(data) {
-	// Split input into two separate arrays of values and weights.
-	const values = data.map((d) => d[0])
-	const weights = data.map((d) => d[1])
-
-	let acc = 0
-	const sum = weights.reduce((acc, element) => acc + element, 0)
-	const weightsSum = weights.map((element) => {
-		acc = element + acc
-		return acc
-	})
-	const rand = Math.random() * sum
-
-	return values[weightsSum.filter((element) => element <= rand).length]
-}
-
 module.exports = {
 	msToTime,
-	format,
 	randint,
 	paginate,
-	pick,
+	loadFiles,
 }
